requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.4
elasticsearch==8.12.1
requests==2.31.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
SQLAlchemy==2.0.27
urllib3==2.2.0
certifi==2024.2.2 